added login route + styles
