+++ quizzer/src/components/Login/login.css
@@ -1,0 +1,53 @@
+.background {
+  height: 800px;
+  width: 1440px;
+  background-color: #f4f6f7;
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+}
+.quizzer {
+  height: 32px;
+  color: #0038fb;
+  font-family: "Open Sans";
+  font-size: 32px;
+  font-weight: bold;
+  line-height: 32px;
+  text-align: center;
+}
+.log-in {
+  height: 32px;
+  width: 312px;
+  color: #000000;
+  font-family: "Open Sans";
+  font-size: 24px;
+  font-weight: bold;
+  line-height: 32px;
+  text-align: center;
+}
+.box {
+  height: 440px;
+  width: 360px;
+  border: 1px solid #e5e5e5;
+  border-radius: 10px;
+  background-color: #ffffff;
+  display: flex;
+  flex-direction: column;
+  align-content: center;
+  justify-content: center;
+  align-items: center;
+}
+.state {
+  background-color: #ea4334;
+}
+.text {
+  height: 44px;
+  width: 216px;
+  color: #ffffff;
+  font-family: "Open Sans";
+  font-size: 18px;
+  font-weight: bold;
+  line-height: 24px;
+}
